Add work replacement feature and fix payroll/leave integration
Key changes:
- Add LeaveReplacement.aspx page for managers to mark leaves as replaced
- Add GetLeavesForWorkReplacement method to LeaveService
- Add SQL migration for IsReplaced, ReplacementDate columns
- Fix payroll generation to calculate leave deduction and OT on creation
- Fix WorkDays calculation: daysInMonth - leaveDays (was hardcoded to 30)
- Fix Admin ID data type: Int32 -> Int16 in EmployeeManagement
- Add work replacement methods: MarkLeaveAsReplaced, CancelWorkReplacement
- Update recalculation to include WorkDays in update fields

Leave replacement workflow:
- Manager marks approved leave as "replaced by work"
- System records replacement date and approver
- Leave quota is returned to employee
- Payroll excludes replaced leaves from deduction calculation

Database migration required:
- Run PHASE7_Migration_01_Leave_Replacement_Columns.sql

+++ Take Time BangPhra/Admin/HR/EmployeeManagement.aspx.cs
@@ -536,7 +536,7 @@ namespace Take_Time_BangPhra.Admin.HR
                     cmd.Parameters.AddWithValue("@LastName", txtLastName.Text.Trim());
                     cmd.Parameters.AddWithValue("@Role", ddlRole.SelectedValue);
 
-                    int newAdminId = Convert.ToInt32(cmd.ExecuteScalar());
+                    short newAdminId = Convert.ToInt16(cmd.ExecuteScalar());
 
                     // Add salary if provided
                     if (salary > 0)
